requirements.txt
streamlit
faiss-cpu
langchain
langchain-community
langchain-openai
openai
python-dotenv
